+++ storage/db-core-delivery/src/main/java/com/phoenix/logistics/storage/db/core/delivery/BaseEntity.java
@@ -1,16 +1,10 @@
-package com.phoenix.logistics.storage.db.core;
+package com.phoenix.logistics.storage.db.core.delivery;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
-
-import java.util.UUID;
+import java.time.LocalDateTime;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
 
 @MappedSuperclass
 public abstract class BaseEntity {
